fix: tests and change database testing to mysql

+++ database/seeders/PermissionSeeder.php
@@ -6,6 +6,7 @@ use App\Enums\PermissionEnum;
 use App\Models\Permission;
 use Illuminate\Database\Seeder;
 use Illuminate\Support\Collection;
+use Illuminate\Support\Facades\App;
 use Illuminate\Support\Facades\Schema;
 use Spatie\Permission\PermissionRegistrar;
 
@@ -20,9 +21,11 @@ class PermissionSeeder extends Seeder
     {
         Collection::make(PermissionEnum::toValues())
             ->tap(function () {
-                Schema::disableForeignKeyConstraints();
-                app(PermissionRegistrar::class)->forgetCachedPermissions();
-                Permission::truncate();
+                if (!App::environment('testing')) {
+                    Schema::disableForeignKeyConstraints();
+                    app(PermissionRegistrar::class)->forgetCachedPermissions();
+                    Permission::truncate();
+                }
             })
             ->each(function ($permission) {
                 Permission::create([
@@ -30,7 +33,9 @@ class PermissionSeeder extends Seeder
                 ]);
             })
             ->tap(function () {
-                Schema::enableForeignKeyConstraints();
+                if (!App::environment('testing')) {
+                    Schema::enableForeignKeyConstraints();
+                }
             });
     }
 }
